+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 2.0-beta-45-SNAPSHOT
- * isssue-71 - Added shape function to main dataset api.  Returns shape in row major 
+ * issue-76 - quartiles for datetime types in descriptive stats
+ * issue-71 - Added shape function to main dataset api.  Returns shape in row major
    format.
  * issue-69 - Columns elide missing values during print operation.